The Merchant Update
Merchants can now be encountered on specific special floors in the
dungeon
Merchants sell 1 weapon from each category (Simple, Martial, Finesse,
Ranged), 1 random piece of armor, 1 bundle of ammunition of a random
type, 1 ration of food, 1 potion of healing, 1 potion of mana, 1 scroll
of identify, 1 scroll of remove curse, and 2 random magic items.

The weapons and armor may be enchanted and may also be upgraded to a
maximum of +4.
Items will be displayed as identified in the merchant interface, and the
player will personally identify an item upon purchasing it.