Items can now generate with upgrades/enchantments
Weapons, Armor, and Shields generated in the dungeon and in chests have
a 25% chance to be upgraded from +1 to +4, a 25% chance to have a random
enchantment, and a 40% chance to be cursed.

Rings generated in the dungeon and in chests have a 40% chance to be
cursed.

Items held by monsters may be enchanted or upgraded, but not cursed,
unless the item in question is dropped by a Mimic, in which case it can
be cursed as normal.

Items generated in structures will never be cursed.

Items sold by Merchants may be enchanted or upgraded, but not cursed.